[Monitor Query] Update metric batch client (#33958)
- Renamed the client to `MetricsClient`.
- Renamed `query_batch` to `query_resources`
- Made `query_resources` positional arguments keyword only
