+++ sdk/monitor/azure-monitor-query/README.md
@@ -19,7 +19,7 @@ The Azure Monitor Query client library is used to execute read-only queries agai
 
 ### Prerequisites
 
-- Python 3.7 or later
+- Python 3.8 or later
 - An [Azure subscription][azure_subscription]
 - A [TokenCredential](https://learn.microsoft.com/python/api/azure-core/azure.core.credentials.tokencredential?view=azure-python) implementation, such as an [Azure Identity library credential type](https://learn.microsoft.com/python/api/overview/azure/identity-readme?view=azure-python#credential-classes).
 - To query Logs, you need one of the following things:
@@ -37,7 +37,7 @@ pip install azure-monitor-query
 
 ### Create the client
 
-An authenticated client is required to query Logs or Metrics. The library includes both synchronous and asynchronous forms of the clients. To authenticate, create an instance of a token credential. Use that instance when creating a `LogsQueryClient`, `MetricsQueryClient`, or `MetricsBatchQueryClient`. The following examples use `DefaultAzureCredential` from the [azure-identity](https://pypi.org/project/azure-identity/) package.
+An authenticated client is required to query Logs or Metrics. The library includes both synchronous and asynchronous forms of the clients. To authenticate, create an instance of a token credential. Use that instance when creating a `LogsQueryClient`, `MetricsQueryClient`, or `MetricsClient`. The following examples use `DefaultAzureCredential` from the [azure-identity](https://pypi.org/project/azure-identity/) package.
 
 #### Synchronous clients
 
@@ -45,11 +45,12 @@ Consider the following example, which creates synchronous clients for both Logs 
 
 ```python
 from azure.identity import DefaultAzureCredential
-from azure.monitor.query import LogsQueryClient, MetricsQueryClient
+from azure.monitor.query import LogsQueryClient, MetricsQueryClient, MetricsClient
 
 credential = DefaultAzureCredential()
-logs_client = LogsQueryClient(credential)
-metrics_client = MetricsQueryClient(credential)
+logs_query_client = LogsQueryClient(credential)
+metrics_query_client = MetricsQueryClient(credential)
+metrics_client = MetricsClient("https://<regional endpoint>", credential)
 ```
 
 #### Asynchronous clients
@@ -58,11 +59,13 @@ The asynchronous forms of the query client APIs are found in the `.aio`-suffixed
 
 ```python
 from azure.identity.aio import DefaultAzureCredential
-from azure.monitor.query.aio import LogsQueryClient, MetricsQueryClient
+from azure.monitor.query.aio import LogsQueryClient, MetricsQueryClient, MetricsClient
 
 credential = DefaultAzureCredential()
-async_logs_client = LogsQueryClient(credential)
-async_metrics_client = MetricsQueryClient(credential)
+async_logs_query_client = LogsQueryClient(credential)
+async_metrics_query_client = MetricsQueryClient(credential)
+async_metrics_client = MetricsClient("https://<regional endpoint>", credential)
+```
 ```
 
 #### Configure clients for non-public Azure clouds
@@ -70,8 +73,8 @@ async_metrics_client = MetricsQueryClient(credential)
 By default, `LogsQueryClient` and `MetricsQueryClient` are configured to connect to the public Azure cloud. These can be configured to connect to non-public Azure clouds by passing in the correct `endpoint` argument: For example:
 
 ```python
-logs_client = LogsQueryClient(credential, endpoint="https://api.loganalytics.azure.cn/v1")
-metrics_client = MetricsQueryClient(credential, endpoint="https://management.chinacloudapi.cn")
+logs_query_client = LogsQueryClient(credential, endpoint="https://api.loganalytics.azure.cn/v1")
+metrics_query_client = MetricsQueryClient(credential, endpoint="https://management.chinacloudapi.cn")
 ```
 
 **Note**: Currently, `MetricsQueryClient` uses the Azure Resource Manager (ARM) endpoint for querying metrics, so you will need the corresponding management endpoint for your cloud when using this client. This is subject to change in the future.
@@ -114,6 +117,7 @@ Each set of metric values is a time series with the following characteristics:
 - [Metrics query](#metrics-query)
   - [Handle metrics query response](#handle-metrics-query-response)
   - [Example of handling response](#example-of-handling-response)
+- [Metrics batch query](#metrics-batch-query)
 
 ### Logs query
 
@@ -527,7 +531,7 @@ for metric in response.metrics:
 
 ### Metrics batch query
 
-A user can also query metrics from multiple resources at once using the `query_batch` method of `MetricsBatchQueryClient`. This uses a different API than the `MetricsQueryClient` and requires that a user pass in a regional endpoint when instantiating the client (for example, "https://westus3.metrics.monitor.azure.com").
+A user can also query metrics from multiple resources at once using the `query_resources` method of `MetricsClient`. This uses a different API than the `MetricsQueryClient` and requires that a user pass in a regional endpoint when instantiating the client (for example, "https://westus3.metrics.monitor.azure.com").
 
 Note, each resource must be in the same region as the endpoint passed in when instantiating the client, and each resource must be in the same Azure subscription. Furthermore, the metric namespace that contains the metrics to be queried must also be passed. A list of metric namespaces can be found [here][metric_namespaces].
 
@@ -538,19 +542,19 @@ import os
 
 from azure.core.exceptions import HttpResponseError
 from azure.identity import DefaultAzureCredential
-from azure.monitor.query import MetricsBatchQueryClient, MetricAggregationType
+from azure.monitor.query import MetricsClient, MetricAggregationType
 
 
 credential = DefaultAzureCredential()
-client = MetricsBatchQueryClient(endpoint, credential)
+client = MetricsClient(endpoint, credential)
 
 resource_uris = [
     "/subscriptions/<id>/resourceGroups/<rg-name>/providers/<source>/storageAccounts/<resource-name-1>",
     "/subscriptions/<id>/resourceGroups/<rg-name>/providers/<source>/storageAccounts/<resource-name-2>"
 ]
 
-response = client.query_batch(
-    resource_uris,
+response = client.query_resources(
+    resource_uris=resource_uris,
     metric_namespace="Microsoft.Storage/storageAccounts",
     metric_names=["Ingress"],
     timespan=timedelta(hours=2),
@@ -586,6 +590,7 @@ The following code samples show common scenarios with the Azure Monitor Query cl
 #### Metrics query samples
 
 - [Send a query using MetricsQueryClient](https://github.com/Azure/azure-sdk-for-python/blob/main/sdk/monitor/azure-monitor-query/samples/sample_metrics_query.py) ([async sample](https://github.com/Azure/azure-sdk-for-python/blob/main/sdk/monitor/azure-monitor-query/samples/async_samples/sample_metrics_query_async.py))
+- [Send a query to multiple resources in a region and subscription using MetricsClient](https://github.com/Azure/azure-sdk-for-python/blob/main/sdk/monitor/azure-monitor-query/samples/sample_metrics_query_multiple.py) ([async sample](https://github.com/Azure/azure-sdk-for-python/blob/main/sdk/monitor/azure-monitor-query/samples/async_samples/sample_metrics_query_multiple_async.py))
 - [Get a list of metric namespaces](https://github.com/Azure/azure-sdk-for-python/blob/main/sdk/monitor/azure-monitor-query/samples/sample_metric_namespaces.py) ([async sample](https://github.com/Azure/azure-sdk-for-python/blob/main/sdk/monitor/azure-monitor-query/samples/async_samples/sample_metric_namespaces_async.py))
 - [Get a list of metric definitions](https://github.com/Azure/azure-sdk-for-python/blob/main/sdk/monitor/azure-monitor-query/samples/sample_metric_definitions.py) ([async sample](https://github.com/Azure/azure-sdk-for-python/blob/main/sdk/monitor/azure-monitor-query/samples/async_samples/sample_metric_definitions_async.py))
