+++ sdk/monitor/azure-monitor-query/assets.json
@@ -2,5 +2,5 @@
   "AssetsRepo": "Azure/azure-sdk-assets",
   "AssetsRepoPrefixPath": "python",
   "TagPrefix": "python/monitor/azure-monitor-query",
-  "Tag": "python/monitor/azure-monitor-query_7462cb1df1"
+  "Tag": "python/monitor/azure-monitor-query_152690fb22"
 }
