+++ sdk/monitor/azure-monitor-query/azure/monitor/query/__init__.py
@@ -6,7 +6,7 @@
 
 from ._logs_query_client import LogsQueryClient
 from ._metrics_query_client import MetricsQueryClient
-from ._metrics_batch_query_client import MetricsBatchQueryClient
+from ._metrics_client import MetricsClient
 
 from ._enums import (
     LogsQueryStatus,
@@ -46,7 +46,7 @@ __all__ = [
     "LogsTableRow",
     "LogsBatchQuery",
     "MetricsQueryClient",
-    "MetricsBatchQueryClient",
+    "MetricsClient",
     "MetricNamespace",
     "MetricNamespaceClassification",
     "MetricDefinition",
