+++ sdk/monitor/azure-monitor-query/azure/monitor/query/_metrics_client.py
@@ -8,31 +8,30 @@ from datetime import timedelta, datetime
 from json import loads
 from typing import Any, List, MutableMapping, Sequence, Optional, Union, Tuple
 
-from azure.core.credentials_async import AsyncTokenCredential
-from azure.core.tracing.decorator_async import distributed_trace_async
+from azure.core.credentials import TokenCredential
+from azure.core.tracing.decorator import distributed_trace
 
-from .._generated.metrics.batch.aio._client import MonitorBatchMetricsClient
-from .._models import MetricsQueryResult
-from ._helpers_async import get_authentication_policy
-from .._helpers import get_timespan_iso8601_endpoints, get_subscription_id_from_resource
+from ._generated.metrics.batch import MonitorBatchMetricsClient
+from ._models import MetricsQueryResult
+from ._helpers import get_authentication_policy, get_timespan_iso8601_endpoints, get_subscription_id_from_resource
 
 JSON = MutableMapping[str, Any]  # pylint: disable=unsubscriptable-object
 
 
-class MetricsBatchQueryClient:  # pylint: disable=client-accepts-api-version-keyword
-    """MetricsBatchQueryClient should be used for performing metrics queries on multiple monitored resources in the
+class MetricsClient:  # pylint: disable=client-accepts-api-version-keyword
+    """MetricsClient should be used for performing metrics queries on multiple monitored resources in the
     same region. A credential with authorization at the subscription level is required when using this client.
 
     :param str endpoint: The regional endpoint to use, for example
         https://eastus.metrics.monitor.azure.com. The region should match the region of the requested
         resources. For global resources, the region should be 'global'. Required.
     :param credential: The credential to authenticate the client.
-    :type credential: ~azure.core.credentials_async.AsyncTokenCredential
+    :type credential: ~azure.core.credentials.TokenCredential
     :keyword str audience: The audience to use when requesting a token. If not provided, the public cloud audience
         will be assumed. Defaults to 'https://metrics.monitor.azure.com'.
     """
 
-    def __init__(self, endpoint: str, credential: AsyncTokenCredential, **kwargs: Any) -> None:
+    def __init__(self, endpoint: str, credential: TokenCredential, **kwargs: Any) -> None:
         self._endpoint = endpoint
         if not self._endpoint.startswith("https://") and not self._endpoint.startswith("http://"):
             self._endpoint = "https://" + self._endpoint
@@ -47,13 +46,13 @@ class MetricsBatchQueryClient:  # pylint: disable=client-accepts-api-version-key
         )
         self._batch_metrics_op = self._client.metrics_batch
 
-    @distributed_trace_async
-    async def query_batch(
+    @distributed_trace
+    def query_resources(
         self,
-        resource_uris: Sequence[str],
-        metric_names: Sequence[str],
-        metric_namespace: str,
         *,
+        resource_uris: Sequence[str],
+        metric_namespace: str,
+        metric_names: Sequence[str],
         timespan: Optional[Union[timedelta, Tuple[datetime, timedelta], Tuple[datetime, datetime]]] = None,
         granularity: Optional[timedelta] = None,
         aggregations: Optional[Sequence[str]] = None,
@@ -65,12 +64,12 @@ class MetricsBatchQueryClient:  # pylint: disable=client-accepts-api-version-key
     ) -> List[MetricsQueryResult]:
         """Lists the metric values for multiple resources.
 
-        :param resource_uris: A list of resource URIs to query metrics for. Required.
-        :type resource_uris: list[str]
-        :param metric_names: The names of the metrics (comma separated) to retrieve. Required.
-        :type metric_names: list[str]
-        :param metric_namespace: Metric namespace that contains the requested metric names. Required.
-        :type metric_namespace: str
+        :keyword resource_uris: A list of resource URIs to query metrics for. Required.
+        :paramtype resource_uris: list[str]
+        :keyword metric_namespace: Metric namespace that contains the requested metric names. Required.
+        :paramtype metric_namespace: str
+        :keyword metric_names: The names of the metrics (comma separated) to retrieve. Required.
+        :paramtype metric_names: list[str]
         :keyword timespan: The timespan for which to query the data. This can be a timedelta,
             a timedelta and a start datetime, or a start datetime/end datetime.
         :paramtype timespan: Optional[Union[~datetime.timedelta, tuple[~datetime.datetime, ~datetime.timedelta],
@@ -81,7 +80,7 @@ class MetricsBatchQueryClient:  # pylint: disable=client-accepts-api-version-key
             `azure.monitor.query.MetricAggregationType` enum to get each aggregation type.
         :paramtype aggregations: Optional[list[str]]
         :keyword max_results: The maximum number of records to retrieve.
-            Valid only if $filter is specified.Defaults to 10.
+            Valid only if $filter is specified. Defaults to 10.
         :paramtype max_results: Optional[int]
         :keyword order_by: The aggregation to use for sorting results and the direction of the sort.
             Only one order can be specified. Examples: sum asc.
@@ -110,9 +109,9 @@ class MetricsBatchQueryClient:  # pylint: disable=client-accepts-api-version-key
 
         .. admonition:: Example:
 
-            .. literalinclude:: ../samples/async_samples/sample_metrics_batch_query_async.py
-                :start-after: [START send_metrics_batch_query_async]
-                :end-before: [END send_metrics_batch_query_async]
+            .. literalinclude:: ../samples/sample_metrics_query_multiple.py
+                :start-after: [START send_metrics_batch_query]
+                :end-before: [END send_metrics_batch_query]
                 :language: python
                 :dedent: 0
                 :caption: Get a response for a batch metrics query.
@@ -127,7 +126,7 @@ class MetricsBatchQueryClient:  # pylint: disable=client-accepts-api-version-key
         resource_id_json: JSON = {"resourceids": list(resource_uris)}
         subscription_id = get_subscription_id_from_resource(resource_uris[0])
 
-        generated = await self._batch_metrics_op.batch(
+        generated = self._batch_metrics_op.batch(
             subscription_id,
             resource_id_json,
             metricnamespace=metric_namespace,
@@ -153,13 +152,13 @@ class MetricsBatchQueryClient:  # pylint: disable=client-accepts-api-version-key
             for value in generated["values"]
         ]
 
-    async def __aenter__(self) -> "MetricsBatchQueryClient":
-        await self._client.__aenter__()
+    def close(self) -> None:
+        """Close the client session."""
+        return self._client.close()
+
+    def __enter__(self) -> "MetricsClient":
+        self._client.__enter__()  # pylint:disable=no-member
         return self
 
-    async def __aexit__(self, *args: Any) -> None:
-        await self._client.__aexit__(*args)
-
-    async def close(self) -> None:
-        """Close the client session."""
-        await self._client.__aexit__()
+    def __exit__(self, *args: Any) -> None:
+        self._client.__exit__(*args)  # pylint:disable=no-member
