+++ sdk/monitor/azure-monitor-query/azure/monitor/query/aio/__init__.py
@@ -6,6 +6,6 @@
 
 from ._logs_query_client_async import LogsQueryClient
 from ._metrics_query_client_async import MetricsQueryClient
-from ._metrics_batch_query_client_async import MetricsBatchQueryClient
+from ._metrics_client_async import MetricsClient
 
-__all__ = ["LogsQueryClient", "MetricsQueryClient", "MetricsBatchQueryClient"]
+__all__ = ["LogsQueryClient", "MetricsQueryClient", "MetricsClient"]
