+++ sdk/monitor/azure-monitor-query/samples/async_samples/sample_metrics_query_multiple_async.py
@@ -1,13 +1,13 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License.
 """
-FILE: sample_metrics_batch_query_async.py
+FILE: sample_metrics_query_multiple_async.py
 DESCRIPTION:
-    This sample demonstrates authenticating the MetricsBatchQueryClient and retrieving the "Ingress"
+    This sample demonstrates authenticating the MetricsClient and retrieving the "Ingress"
     metric along with the "Average" aggregation type for multiple resources.
     The query will execute over a timespan of 2 hours with a granularity of 5 minutes.
 USAGE:
-    python sample_metrics_batch_query_async.py
+    python sample_metrics_query_multiple_async.py
     1) AZURE_METRICS_ENDPOINT - The regional metrics endpoint to use (i.e. https://westus3.metrics.monitor.azure.com)
 
     This example uses DefaultAzureCredential, which requests a token from Azure Active Directory.
@@ -24,14 +24,14 @@ import os
 from azure.core.exceptions import HttpResponseError
 from azure.identity.aio import DefaultAzureCredential
 from azure.monitor.query import MetricAggregationType
-from azure.monitor.query.aio import MetricsBatchQueryClient
+from azure.monitor.query.aio import MetricsClient
 
 
 async def query_metrics_batch():
     endpoint = os.environ["AZURE_METRICS_ENDPOINT"]
 
     credential = DefaultAzureCredential()
-    client = MetricsBatchQueryClient(endpoint, credential)
+    client = MetricsClient(endpoint, credential)
 
     resource_uris = [
         '/subscriptions/<id>/resourceGroups/<rg>/providers/Microsoft.Storage/storageAccounts/<account-1>',
@@ -39,10 +39,10 @@ async def query_metrics_batch():
     ]
     async with client:
         try:
-            response = await client.query_batch(
+            response = await client.query_resources(
                 resource_uris=resource_uris,
-                metric_names=["Ingress"],
                 metric_namespace="Microsoft.Storage/storageAccounts",
+                metric_names=["Ingress"],
                 timespan=timedelta(hours=2),
                 granularity=timedelta(minutes=5),
                 aggregations=[MetricAggregationType.AVERAGE],
